Feat: Req-84 ao Req-87
