+++ src/styles/Profile.css
@@ -1,5 +1,5 @@
 .profile-btns {
-  margin-top: 200px;
+  margin-top: 150px;
 }
 
 button {
